- Fix du chat 5.0.5 - Correction des flags positive/negative sur les Aura

+++ src/server/game/Battlegrounds/Battleground.cpp
@@ -1720,7 +1720,7 @@ void Battleground::SendWarningToAll(int32 entry, ...)
     data << (uint64)0;
     data << (uint32)(msg.length() + 1);
     data << msg.c_str();
-    data << (uint8)0;
+    data << (uint16)0;
     data << (float)0.0f;                                   // added in 4.2.0, unk
     data << (uint8)0;                                      // added in 4.2.0, unk
     for (BattlegroundPlayerMap::const_iterator itr = m_Players.begin(); itr != m_Players.end(); ++itr)
