+++ src/server/game/Chat/Channels/Channel.cpp
@@ -653,7 +653,7 @@ void Channel::Say(uint64 p, const char *what, uint32 lang)
         data << p;
         data << messageLength;
         data << what;
-        data << uint8(player ? player->GetChatTag() : 0);
+        data << uint16(player ? player->GetChatTag() : 0);
 
         SendToAll(&data, !players[p].IsModerator() ? p : false);
     }
